SUGGESTION Remove the space before the opening parenthesis. This generates the following warning in some pep8 tools:
report.py:26:10: E211 whitespace before '('
Not all pep8 tools generate this warning, so this is just a suggestion. In Python 3, print is a full fledged function, so there should be no space before the parenthesis. Even if you are using Python 2, eliminating the space will get you into good habits.