SUGGESTION If a user was unfamiliar with this project, they won't know what you are referring to by "the database". Be more descriptive. Discuss what database you are talking about, and what the purpose of that database is. You could mention that it is a mock database for a news site. You could briefly mention the type of data the database contains, and what questions your report answers.
When writing your README files, assume you are writing for a general audience that is unfamiliar with this project. This will give you practice writing robust, easy to follow README files. This is a valuable skill for a developer.

These projects can be a valuable part of your portfolio. If you want to apply for a job as a developer, the people doing the hiring will first look at your README files. So you want to clearly communicate in your README files what the purpose of your project is so they will be interested to learn more about you. I recommend writing your README files for these projects with future potential employers in mind.

REQUIRED
A reader who was not familiar with this course would not know what fullstack-nanodegree-vm-master is. You could provide a link to it. However there should be no need for the user to download the entire fullstack repository. If you want the user to use Vagrant, then I would just include a Vagrantfile in your repository. Include explicit directions on how to power up and log into the virtual machine if you wish the user to use a virtual machine. For example, you could provide directions such as:

From the command line, navigate to the folder containing the Vagrantfile
Power up the virtual machine by typing: vagrant up note: this may take a couple minutes to complete
Once the virtual machine is done booting, log into it by typing: vagrant ssh
Also, when I ran your script by following your directions, I get the following error:

Traceback (most recent call last):
  File "report.py", line 10, in <module>
    conn = psycopg2.connect("dbname='news'")
  File "/usr/local/lib/python2.7/dist-packages/psycopg2/__init__.py", line 130, in connect
    conn = _connect(dsn, connection_factory=connection_factory, **kwasync)
psycopg2.OperationalError: FATAL:  database "news" does not exist
You are missing a few critical steps. You need to provide explicit instructions on how to set up the database. Provide step by step instructions and assume the user is starting from scratch. Some steps you may want to include:

how to create the news database
where to get the newsdata.sql file with the database schema and data.
you could include a compressed (e.g. zipped) version of newsdata.sql in your repository, or include a link to where the user can download this file.
how to import the schema and data into the database